backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
google-generativeai
pillow
python-multipart
requests
beautifulsoup4